fix(st-02003): address review comments on insert tooling
